Add batch processing pipeline to assignment file
This commit adds the batch processing pipeline after the single image
processing pipeline. The batch processing now computes and displays all
intermediary results used in the single image processing pipeline to
justify the "magic numbers".

Additionally, the `sigmaX` value for the Gaussian blur is now
dynamically computed using the normal distribution parameters of the
original image and the observations made from the histogram of the
orignal image.

NOTES:

- For some unknown reason, the processed images are NOT displayed in
grayscale.